[feature] externalização das funções do filtro
