+++ src/app/app.component.ts
@@ -1,9 +1,8 @@
-import { Component, OnInit } from '@angular/core';
 import { IUser } from './interfaces/user/user.interface';
 import { UsersList } from './data/users-list';
 import { IFilterOptions } from './interfaces/filter-options.interface';
-import { IStatus } from './interfaces/user/status.interface';
-import { isWithinInterval } from 'date-fns';
+import { filterUsersList } from './utils/filter-users-list';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -30,51 +29,7 @@ export class AppComponent implements OnInit {
   }
 
   onFilter(filterOptions: IFilterOptions) {
-    this.userListFiltered = this.filterUsersList(filterOptions, this.usersList);
-  }
-  filterUsersList(filterOptions: IFilterOptions, usersList: IUser[]): IUser[] {
-    let filteredList: IUser[] = [];
-
-    filteredList = this.filterByName(filterOptions.name, usersList);
-    filteredList = this.filterByStatus(filterOptions.status, filteredList);
-    filteredList = this.filterByDate(filterOptions.startDate, filterOptions.endDate, filteredList);
-
-
-    return filteredList;
+    this.userListFiltered = filterUsersList(filterOptions, this.usersList);
   }
 
-  filterByDate(startDate: Date | undefined, endDate: Date | undefined, userslist: IUser[]): IUser[] {
-    const DATES_NOT_SELECTED = startDate === undefined || endDate === undefined;
-
-    if (DATES_NOT_SELECTED)
-      return userslist;
-
-    const listFiltered = userslist.filter((user) =>
-      isWithinInterval(new Date(user.dataCadastro), { start: startDate, end: endDate }));
-
-    return listFiltered;
-
-  }
-
-  filterByStatus(status: boolean | undefined, filteredList: IUser[]): IUser[] {
-    const STATUS_NOT_SELECTED = status === undefined;
-
-    if (STATUS_NOT_SELECTED)
-      return filteredList;
-
-    const filteredListStatus = filteredList.filter((user) => user.ativo === status);
-
-    return filteredListStatus;
-  }
-
-  filterByName(name: string | undefined, usersList: IUser[]): IUser[] {
-    const NAME_NOT_TYPPED = name === undefined;
-
-    if (NAME_NOT_TYPPED)
-      return usersList;
-
-    const filteredList = usersList.filter((user) => user.nome.toLowerCase().includes(name.toLowerCase()));
-
-    return filteredList;
-  }
 }
